+++ configs/llama_60m.json
@@ -1,0 +1,20 @@
+{
+    "architectures": [
+        "LLaMAForCausalLM"
+    ],
+    "bos_token_id": 0,
+    "eos_token_id": 1,
+    "hidden_act": "silu",
+    "hidden_size": 512,
+    "intermediate_size": 1376,
+    "initializer_range": 0.02,
+    "max_sequence_length": 1024,
+    "model_type": "llama",
+    "num_attention_heads": 8,
+    "num_hidden_layers": 8,
+    "pad_token_id": -1,
+    "rms_norm_eps": 1e-06,
+    "transformers_version": "4.28.1",
+    "use_cache": false,
+    "vocab_size": 32000
+}
