+++ src/hyperatt/main.py
@@ -1,5 +1,4 @@
 import random
-import argparse
 import numpy as np
 
 import torch
@@ -12,59 +11,12 @@ import datasets
 
 from loguru import logger
 
-import training_utils
 from dataloader import PreprocessedIterableDataset
 from llama import LlamaForCausalLM
 from trainer import train_model
+from utils import parse_args
 
 transformers.logging.set_verbosity_error()
-
-def parse_args(args):
-    parser = argparse.ArgumentParser()
-
-    parser.add_argument("--model_config", type=str, required=True)
-    parser.add_argument("--use_hf_model", default=False, action="store_true")
-    parser.add_argument("--continue_from", type=str, default=None)
-    parser.add_argument("--batch_size", type=int, required=True)
-    parser.add_argument("--gradient_accumulation", type=int, default=None)
-    parser.add_argument("--total_batch_size", type=int, default=None)
-    parser.add_argument("--max_length", type=int, default=256)
-    parser.add_argument("--optimizer", default="Adam")
-    parser.add_argument("--lr", type=float, default=1e-4)
-    parser.add_argument("--scheduler", type=str, default="cosine", choices=["linear", "cosine", "cosine_restarts"])
-    parser.add_argument("--min_lr_ratio", type=float, default=0.1)
-    parser.add_argument("--activation_checkpointing", action="store_true")
-    parser.add_argument("--weight_decay", type=float, default=0.0)
-    parser.add_argument("--warmup_steps", type=int, default=1_000)
-    parser.add_argument("--eval_every", type=int, default=5_000)
-    parser.add_argument("--num_training_steps", type=int, default=10_000,
-                        help="Number of **update steps** to train for. "
-                             "Notice that gradient accumulation is taken into account.")
-    parser.add_argument("--max_train_tokens", type=training_utils.max_train_tokens_to_number, default=None,
-                        help="Number of tokens to train on. Overwrites num_training_steps. "
-                             "You can use M and B suffixes, e.g. 100M or 1B.")
-    parser.add_argument("--save_every", type=int, default=10_000)
-    parser.add_argument("--save_dir", type=str, default=None)
-    parser.add_argument("--tags", type=str, default=None)
-    parser.add_argument("--dtype", type=str, default="bfloat16" if torch.cuda.is_bf16_supported() else "float32")
-    parser.add_argument("--workers", type=int, default=8)
-    parser.add_argument("--seed", type=int, default=0)
-    parser.add_argument("--name", type=str, default="test")
-    parser.add_argument("--grad_clipping", type=float, default=0.0)   
-    # beta1 for adafactor
-    parser.add_argument("--beta1", type=float, default=0.0)
-    
-    # GaLore parameters
-    parser.add_argument("--rank", type=int, default=128)
-    parser.add_argument("--update_proj_gap", type=int, default=50)
-    parser.add_argument("--galore_scale", type=float, default=1.0)
-    parser.add_argument("--proj_type", type=str, default="std")
-    
-    args = parser.parse_args(args)
-
-    args = args_utils.check_args_torchrun_main(args)
-    return args
-
 
 
 def main(args):
@@ -108,14 +60,11 @@ def main(args):
     dataset = PreprocessedIterableDataset(data, tokenizer, batch_size=args.batch_size, max_length=args.max_length)
     dataloader = torch.utils.data.DataLoader(dataset, batch_size=None, num_workers=args.workers)
 
-    model_config = AutoConfig.from_pretrained(args.model_config)
+    model_config = AutoConfig.from_pretrained("./configs/"+args.model_config +".json")
     model = LlamaForCausalLM(model_config)
 
     if args.activation_checkpointing:
         model.gradient_checkpointing_enable()
-
-
-    
 
     train_model(model, model_config, dataloader, device, args)
