+++ src/hyperatt/utils.py
@@ -1,39 +1,81 @@
-from typing import Dict, Tuple
-from transformers import TrainingArguments, HfArgumentParser
+import os
 import argparse
-from transformers.hf_argparser import DataClass
+from datetime import datetime
+from loguru import logger
+import training_utils
+import torch
 
-import dataclasses
-from typing import Tuple, Union, Iterable, NewType, Any
 
-DataClassType = NewType("DataClassType", Any)
+def parse_args(args):
+    parser = argparse.ArgumentParser()
 
-def parse_args(file_path):
-    # This is going to need a lot of work to fix tbh
-    argp = CustomArgumentParser(TrainingArguments)
-    training_args, remianing_args = argp.parse_json_file(file_path)
+    parser.add_argument("--model_config", type=str, default='llama_60m')
+    parser.add_argument("--use_hf_model", default=False, action="store_true")
+    parser.add_argument("--continue_from", type=str, default=None)
+    parser.add_argument("--batch_size", type=int, default=256)
+    parser.add_argument("--gradient_accumulation", type=int, default=None)
+    parser.add_argument("--total_batch_size", type=int, default=512)
+    parser.add_argument("--max_length", type=int, default=256)
+    parser.add_argument("--optimizer", default="galore_adamw")
+    parser.add_argument("--lr", type=float, default=0.01)
+    parser.add_argument("--scheduler", type=str, default="cosine", choices=["linear", "cosine", "cosine_restarts"])
+    parser.add_argument("--min_lr_ratio", type=float, default=0.1)
+    parser.add_argument("--activation_checkpointing", action="store_true")
+    parser.add_argument("--weight_decay", type=float, default=0.0)
+    parser.add_argument("--warmup_steps", type=int, default=1_000)
+    parser.add_argument("--do_eval", action='store_false', default=True)
+    parser.add_argument("--eval_every", type=int, default=1_000)
+    parser.add_argument("--num_training_steps", type=int, default=10_000,
+                        help="Number of **update steps** to train for. "
+                             "Notice that gradient accumulation is taken into account.")
+    parser.add_argument("--max_train_tokens", type=training_utils.max_train_tokens_to_number, default=None,
+                        help="Number of tokens to train on. Overwrites num_training_steps. "
+                             "You can use M and B suffixes, e.g. 100M or 1B.")
+    parser.add_argument("--save_every", type=int, default=10_000)
+    parser.add_argument("--save_dir", type=str, default=None)
+    parser.add_argument("--tags", type=str, default=None)
+    parser.add_argument("--dtype", type=str, default="bfloat16" if torch.cuda.is_bf16_supported() else "float32")
+    parser.add_argument("--workers", type=int, default=8)
+    parser.add_argument("--seed", type=int, default=0)
+    parser.add_argument("--name", type=str, default="test")
+    parser.add_argument("--grad_clipping", type=float, default=0.0)   
+    # beta1 for adafactor
+    parser.add_argument("--beta1", type=float, default=0.0)
     
-    args = argparse.Namespace(**remianing_args)
-
-    return training_args, args
-
-
-class CustomArgumentParser(HfArgumentParser):
-    def __init__(self, dataclass_types: Union[DataClassType, Iterable[DataClassType]], **kwargs):
-        super().__init__(dataclass_types, **kwargs)
+    # GaLore parameters
+    parser.add_argument("--rank", type=int, default=128)
+    parser.add_argument("--update_proj_gap", type=int, default=200)
+    parser.add_argument("--galore_scale", type=float, default=0.25)
+    parser.add_argument("--proj_type", type=str, default="std")
     
-    def parse_dict(self, args: Dict[str, Any], allow_extra_keys: bool = False) -> Tuple[DataClass]:
-        unused_keys = set(args.keys())
-        outputs = []
-        
-        for dtype in self.dataclass_types:
-            keys = {f.name for f in dataclasses.fields(dtype) if f.init}
-            #print(keys)
-            inputs = {k: v for k,v in args.items() if k in keys}
-            unused_keys.difference_update(inputs.keys())
-            obj = dtype(**inputs)
-            outputs.append(obj)
-        
-        remaining_dict = {k: v for k, v in args.items() if k in unused_keys}
+    args = parser.parse_args(args)
 
-        return (*outputs, remaining_dict)
+    args = check_args_torchrun_main(args)
+    return args
+
+def check_args_torchrun_main(args):
+
+    if args.save_dir is None:
+        # use checkpoints / model name, date and time as save directory
+        args.save_dir = f"checkpoints/{args.model_config.split('/')[-1].rstrip('.json')}-{datetime.now().strftime('%Y-%m-%d-%H-%M-%S')}"
+
+    if args.tags is not None:
+        args.tags = args.tags.split(",")
+
+    if args.total_batch_size is None:
+        args.gradient_accumulation = args.gradient_accumulation or 1
+        args.total_batch_size = args.batch_size * args.gradient_accumulation
+
+    assert args.total_batch_size % args.batch_size == 0, "total_batch_size must be divisible by batch_size"
+
+    if args.max_train_tokens is not None:
+        args.num_training_steps = args.max_train_tokens // args.total_batch_size
+        logger.info(f"Training for {args.num_training_steps} update steps")
+
+    if args.continue_from is not None:
+        assert os.path.exists(args.continue_from), f"--continue_from={args.continue_from} does not exist"
+
+    if args.dtype in ["fp16", "float16"]:
+        raise NotImplementedError("fp16 is not supported in torchrun_main.py. Use deepspeed_main.py instead (but it seems to have bugs)")
+
+    return args
